requirements.txt
ecdsa
pycrypto
fabric
tornado
redis
python-twitter
oauth2
pyopenssl